+++ es6/docxtemplater.js
@@ -24,7 +24,7 @@ const {
 	throwApiVersionError,
 } = require("./errors.js");
 
-const currentModuleApiVersion = [3, 33, 0];
+const currentModuleApiVersion = [3, 34, 0];
 
 const Docxtemplater = class Docxtemplater {
 	constructor(zip, { modules = [], ...options } = {}) {