+++ es6/tests/e2e/base.js
@@ -111,7 +111,7 @@ describe("Api versioning", function () {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 33, 0],
+					currentModuleApiVersion: [3, 34, 0],
 					neededVersion: [5, 6, 0],
 				},
 			}
@@ -126,14 +126,14 @@ describe("Api versioning", function () {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 33, 0],
+					currentModuleApiVersion: [3, 34, 0],
 					neededVersion: [3, 44, 0],
 				},
 			}
 		);
 
 		expectToThrow(
-			doc.verifyApiVersion.bind(null, "3.33.100"),
+			doc.verifyApiVersion.bind(null, "3.34.100"),
 			Errors.XTAPIVersionError,
 			{
 				message:
@@ -141,8 +141,8 @@ describe("Api versioning", function () {
 				name: "APIVersionError",
 				properties: {
 					id: "api_version_error",
-					currentModuleApiVersion: [3, 33, 0],
-					neededVersion: [3, 33, 100],
+					currentModuleApiVersion: [3, 34, 0],
+					neededVersion: [3, 34, 100],
 				},
 			}
 		);